# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AlphaPilot API",
    "description": "AI Autonomous Trading System",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/positions": {
      "get": {
        "tags": [
          "positions"
        ],
        "summary": "List Positions",
        "description": "列出所有开仓持仓 (要求登录)。",
        "operationId": "list_positions",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "ap_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ap Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response_list_PositionRead__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/positions/{position_id}/sltp": {
      "patch": {
        "tags": [
          "positions"
        ],
        "summary": "Update Position Sltp",
        "description": "修改持仓 SL/TP (handoff P2 §3.2): 走守卫规则校验 + 审计。\n\n权限: 暂 require_admin, P4 RBAC 落地后放宽为 trader+。",
        "operationId": "update_position_sltp",
        "parameters": [
          {
            "name": "position_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Position Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "ap_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ap Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SltpUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response_SltpOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/trades": {
      "get": {
        "tags": [
          "trades"
        ],
        "summary": "List Trades",
        "description": "返回最近 N 条已完成交易记录 (要求登录)。",
        "operationId": "list_trades",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "ap_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ap Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response_list_TradeRead__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/account": {
      "get": {
        "tags": [
          "account"
        ],
        "summary": "Get Account",
        "description": "返回最新账户快照 (要求登录)。",
        "operationId": "get_account",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "ap_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ap Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response_AccountSnapshotRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/orders/precheck": {
      "post": {
        "tags": [
          "orders"
        ],
        "summary": "Precheck Order",
        "description": "守卫预检: 逐项 PASS/FAIL + 总判定 (只读, 下单时服务端会再跑一遍)。",
        "operationId": "precheck_order",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "ap_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ap Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualOrderCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response_PrecheckOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/orders": {
      "post": {
        "tags": [
          "orders"
        ],
        "summary": "Place Order",
        "description": "手动下单: 服务端重跑守卫, HALTED 仅接受 reduce-only 平仓, 幂等 + 审计。",
        "operationId": "place_order",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "ap_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ap Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualOrderCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response_OrderPlacedOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ManualOrderCreate": {
        "properties": {
          "symbol": {
            "type": "string",
            "maxLength": 20,
            "minLength": 1,
            "title": "Symbol"
          },
          "side": {
            "type": "string",
            "enum": [
              "BUY",
              "SELL"
            ],
            "title": "Side"
          },
          "type": {
            "type": "string",
            "enum": [
              "MARKET",
              "LIMIT"
            ],
            "title": "Type",
            "default": "MARKET"
          },
          "qty": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Qty"
          },
          "price": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "sl": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Sl"
          },
          "tp": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Tp"
          },
          "reduce_only": {
            "type": "boolean",
            "title": "Reduce Only",
            "default": false
          },
          "client_order_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 40,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Order Id"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "side",
          "qty"
        ],
        "title": "ManualOrderCreate",
        "description": "POST /api/orders 与 /api/orders/precheck 共用入参。"
      },
...
      "PositionRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "quantity": {
            "type": "number",
            "title": "Quantity"
          },
          "entry_price": {
            "type": "number",
            "title": "Entry Price"
          },
          "current_price": {
            "type": "number",
            "title": "Current Price"
          },
          "stop_loss": {
            "type": "number",
            "title": "Stop Loss"
          },
          "take_profit": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Take Profit"
          },
          "unrealized_pnl": {
            "type": "number",
            "title": "Unrealized Pnl"
          },
          "unrealized_pnl_pct": {
            "type": "number",
            "title": "Unrealized Pnl Pct"
          },
          "opened_at": {
            "type": "string",
            "title": "Opened At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "symbol",
          "quantity",
          "entry_price",
          "current_price",
          "stop_loss",
          "unrealized_pnl",
          "unrealized_pnl_pct",
          "opened_at"
        ],
        "title": "PositionRead"
      },
      "PrecheckOut": {
        "properties": {
          "verdict": {
            "type": "string",
            "title": "Verdict"
          },
          "halted": {
            "type": "boolean",
            "title": "Halted"
          },
          "checks": {
            "items": {
              "$ref": "#/components/schemas/GuardCheckItem"
            },
            "type": "array",
            "title": "Checks"
          },
          "context": {
            "additionalProperties": true,
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "verdict",
          "halted",
          "checks",
          "context"
        ],
        "title": "PrecheckOut"
      },
...
      "Response_AccountSnapshotRead_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "code": {
            "type": "string",
            "title": "Code",
            "default": "0"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "成功"
          },
          "detailMessage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detailmessage"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AccountSnapshotRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "type": "object",
        "title": "Response[AccountSnapshotRead]"
      },
...
      "Response_OrderPlacedOut_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "code": {
            "type": "string",
            "title": "Code",
            "default": "0"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "成功"
          },
          "detailMessage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detailmessage"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OrderPlacedOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "type": "object",
        "title": "Response[OrderPlacedOut]"
      },
      "Response_PrecheckOut_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "code": {
            "type": "string",
            "title": "Code",
            "default": "0"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "成功"
          },
          "detailMessage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detailmessage"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PrecheckOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "type": "object",
        "title": "Response[PrecheckOut]"
      },
...
      "Response_SltpOut_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "code": {
            "type": "string",
            "title": "Code",
            "default": "0"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "成功"
          },
          "detailMessage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detailmessage"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SltpOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "type": "object",
        "title": "Response[SltpOut]"
      },
...
      "Response_list_PositionRead__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "code": {
            "type": "string",
            "title": "Code",
            "default": "0"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "成功"
          },
          "detailMessage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detailmessage"
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/PositionRead"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "type": "object",
        "title": "Response[list[PositionRead]]"
      },
...
      "Response_list_TradeRead__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "code": {
            "type": "string",
            "title": "Code",
            "default": "0"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "成功"
          },
          "detailMessage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detailmessage"
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/TradeRead"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "type": "object",
        "title": "Response[list[TradeRead]]"
      },
...
      "SltpOut": {
        "properties": {
          "position_id": {
            "type": "integer",
            "title": "Position Id"
          },
          "stop_loss": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stop Loss"
          },
          "take_profit": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Take Profit"
          }
        },
        "type": "object",
        "required": [
          "position_id"
        ],
        "title": "SltpOut"
      },
      "SltpUpdate": {
        "properties": {
          "stop_loss": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Stop Loss"
          },
          "take_profit": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Take Profit"
          }
        },
        "type": "object",
        "title": "SltpUpdate",
        "description": "PATCH /api/positions/{id}/sltp 入参; 至少给一个字段。"
      },
...
      "TradeRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "quantity": {
            "type": "number",
            "title": "Quantity"
          },
          "entry_price": {
            "type": "number",
            "title": "Entry Price"
          },
          "exit_price": {
            "type": "number",
            "title": "Exit Price"
          },
          "pnl": {
            "type": "number",
            "title": "Pnl"
          },
          "pnl_pct": {
            "type": "number",
            "title": "Pnl Pct"
          },
          "exit_reason": {
            "type": "string",
            "title": "Exit Reason"
          },
          "strategy_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strategy Mode"
          },
          "regime": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Regime"
          },
          "opened_at": {
            "type": "string",
            "title": "Opened At"
          },
          "closed_at": {
            "type": "string",
            "title": "Closed At"
          },
          "holding_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Holding Seconds"
          }
        },
        "type": "object",
        "required": [
          "id",
          "symbol",
          "quantity",
          "entry_price",
          "exit_price",
          "pnl",
          "pnl_pct",
          "exit_reason",
          "opened_at",
          "closed_at"
        ],
        "title": "TradeRead"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}